pyproject.toml
[project]
name = "exam-project"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "kaggle>=1.7.4.5",
    "matplotlib>=3.10.7",
    "mlflow-skinny>=3.6.0",
    "pandas>=2.3.3",
    "pandas-stubs==2.3.2.250926",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
    "torch>=2.9.1",
]
